src/control-catalog/behave_tests/requirements.txt
behave==1.2.6
black==24.3.0
boto3==1.34.78
botocore==1.34.78
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-kms==2.21.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pathspec==0.12.1
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
